loading model and creating prediction engine

+++ PricePrediction/ModelOutput.cs
@@ -3,6 +3,6 @@ namespace PricePrediction
 {
     public class ModelOutput
     {
-        public double Score { get; set; }
+        public float Score { get; set; }
     }
 }
